added "dry run" option to Modification-spewing methods
Now there is a way to tell to methods producing Modification (typically, those of "actual" Block) *not* to do the changes. The use of this is if the method has to modify both the physical and the abstract representation, but the abstract one is already changed. This happens when reacting to an abstract modification.