changed handling in the state of a [Col]Variable
Variable is now explicitly fixed and unfixed by setting the LSB of f_state. The other bits are available to derived classes. ColVariable uses them to keep the "type" (binary, int, continuous).

Also, VariableMod went against the 7 golden rules of Modification by storing the state that the Variable was set to; this is now fixed.