Clarified methods/functions confusion in the comments
In order to make the comments clearer, (almost) all mentions to "method" were
removed. When appropriate, mentions to "class member functions" were
added. Also, "signature" was removed from the comments, as this term was not
being used precisely, and FunctionSignature and MethodSignature were renamed to
FunctionType and MemberFunctionType, respectively. The name "methods factory"
and the names of the functions still remain the same.