Added support for boost::multi_array
It still remains to implement the case in which a boost multi_array is allowed
to be serialized with dimensions whose sizes are different from those provided
by the ncDim parameter. This is the case, for example, of a two-dimensional
array that can be represented as a one-dimensional array (the values do not
vary across the other dimension).