Added BlockSolverConfig, RBlockSolverConfig, and ERBlockSolverConfig
BlockSolverConfig was divided into three classes: BlockSolverConfig,
RBlockSolverConfig, and ERBlockSolverConfig. The first one considers the
configuration of the Solver of the Block only, the second one also considers
the configuration of the Solver of the sub-Block, and the last one also
consider the configuration of the Solver of the "indirect sub-Block" (those
that may be part of the Constraint and Objective).