rather large changes in LagBFunction
Mostly, CostMatrix now has vector of pairs < index of y variable , coeff > rather than < pointer to y variable , coeff > like before. This seems to be better almost everywhere. The trigger was a rather catastrophic conceptual error while getting the matrix A by columns, where the order of the columns was all wrong. That part has been restructured and it should be better (but tests are still in progress).