fixed particularly nasty compiler-dependent bugs
Some pieces of codes in LagBFunction and PolyhedralFunction contained expressions where some parts had side effect on others (which.empty() and std::move( which )), which meant that the results were only correct depending on assumptions on the order in which the pieces were executed, which is not fixed in C++ and is compiler dependent. Of course this only reared its ugly head in the worst possible situation.